# Liquidation: increase number of samples for small miners